(+semver: breaking) Fix implementations of View and Source (especially View.Keys()) (#37)
* (+semver: breaking) Fix implementation of View.Keys()

This requires a change to the return value from Keys() in order to pass through a possible error from retrieving the iterator from the view.

closes #36
closes #30 

* (+semver: fix) Wait for underlying PartitionTable to become ready before attempting operations on the table

This does not change the programming interfaces, but does represent a nontrivial change in behavior. ViewSources and ViewSinks may trigger at longer-than-expected intervals, depending on the readiness of the PartitionTable. (Readiness is evaluated on each "tick" of the timer.)

Closes #38

* (+semver: breaking) Removed Watch() func from Kafmesh View and Source; made API parallel to other Kafmesh objects

Now all of the New_*_{Object} funcs return funcs to pass to the Service.RegisterRunner() func. Also, Kafmesh Views and Sources include a context that is bound to the context under which they are running.

Refactored building contexts to more clearly represent how they are related to one another.

Also replaced invalid "WaitRecovered" func with proper "WaitRunning" func

---

Relates to internal issue EN-7623